Add cleanup command
This also refacctors some common code to common.go
