+++ gitpod-network-check/cmd/checks.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
@@ -21,28 +20,6 @@ import (
 	"github.com/spf13/cobra"
 	"golang.org/x/sync/errgroup"
 	"k8s.io/apimachinery/pkg/util/wait"
-)
-
-const gitpodRoleName = "GitpodNetworkCheck"
-const gitpodInstanceProfile = "GitpodNetworkCheck"
-
-var networkCheckTag = []iam_types.Tag{
-	{
-		Key:   aws.String("gitpod.io/network-check"),
-		Value: aws.String("true"),
-	},
-}
-
-func initAwsConfig(ctx context.Context, region string) (aws.Config, error) {
-	return config.LoadDefaultConfig(ctx, config.WithRegion(region))
-}
-
-// this will be useful when we are cleaning up things at the end
-var (
-	InstanceIds     []string
-	SecurityGroups  []string
-	Roles           []string
-	InstanceProfile string
 )
 
 var checkCommand = &cobra.Command{ // nolint:gochecknoglobals
@@ -272,7 +249,7 @@ func launchInstances(ctx context.Context, ec2Client *ec2.Client, subnets []strin
 	for _, subnet := range subnets {
 		secGroup, err := createSecurityGroups(ctx, ec2Client, subnet)
 		if err != nil {
-			return nil, fmt.Errorf("❌ failed to create security group: %v", err)
+			return nil, fmt.Errorf("❌ failed to create security group for subnet '%v': %v", subnet, err)
 		}
 		SecurityGroups = append(SecurityGroups, secGroup)
 		instanceId, err := launchInstanceInSubnet(ctx, ec2Client, subnet, secGroup, profileArn)
@@ -484,64 +461,6 @@ func createSecurityGroups(ctx context.Context, svc *ec2.Client, subnetID string)
 	}
 
 	return *sgID, nil
-}
-
-func cleanup(ctx context.Context, svc *ec2.Client, iamsvc *iam.Client) {
-	if len(InstanceIds) > 0 {
-		_, err := svc.TerminateInstances(ctx, &ec2.TerminateInstancesInput{
-			InstanceIds: InstanceIds,
-		})
-		if err != nil {
-			log.WithError(err).WithField("instanceIds", InstanceIds).Warnf("Failed to cleanup instances, please cleanup manually")
-		}
-	}
-	if len(Roles) > 0 {
-		for _, role := range Roles {
-			_, err := iamsvc.DetachRolePolicy(ctx, &iam.DetachRolePolicyInput{PolicyArn: aws.String("arn:aws:iam::aws:policy/AmazonSSMManagedInstanceCore"), RoleName: aws.String(role)})
-			if err != nil {
-				log.WithError(err).WithField("rolename", role).Warnf("Failed to cleanup role, please cleanup manually")
-			}
-
-			_, err = iamsvc.RemoveRoleFromInstanceProfile(ctx, &iam.RemoveRoleFromInstanceProfileInput{
-				RoleName:            aws.String(role),
-				InstanceProfileName: aws.String(InstanceProfile),
-			})
-			if err != nil {
-				log.WithError(err).WithField("roleName", role).WithField("profileName", InstanceProfile).Warnf("Failed to remove role from instance profile")
-			}
-
-			_, err = iamsvc.DeleteRole(ctx, &iam.DeleteRoleInput{RoleName: aws.String(role)})
-			if err != nil {
-				log.WithError(err).WithField("rolename", role).Warnf("Failed to cleanup role, please cleanup manaullay")
-			}
-		}
-
-		_, err := iamsvc.DeleteInstanceProfile(ctx, &iam.DeleteInstanceProfileInput{
-			InstanceProfileName: aws.String(InstanceProfile),
-		})
-
-		if err != nil {
-			log.WithError(err).WithField("instanceProfile", InstanceProfile).Warnf("Failed to clean up instance profile, please cleanup manually")
-		}
-	}
-
-	log.Info("Cleaning up: Waiting for 1 minute so network interfaces are deleted")
-	time.Sleep(time.Minute)
-
-	if len(SecurityGroups) > 0 {
-		for _, sg := range SecurityGroups {
-			deleteSGInput := &ec2.DeleteSecurityGroupInput{
-				GroupId: aws.String(sg),
-			}
-
-			_, err := svc.DeleteSecurityGroup(ctx, deleteSGInput)
-			if err != nil {
-				log.WithError(err).WithField("securityGroup", sg).Warnf("Failed to clean up security group, please cleanup manually")
-			}
-
-		}
-
-	}
 }
 
 func createIAMRoleAndAttachPolicy(ctx context.Context, svc *iam.Client) (*iam_types.Role, error) {
